Fast-fail tests if there are problems in the initial load.
Rather than wait for $P to be present in the page for the full timeout
of a test, wait for the document to enter the 'complete' readyState. At
this point, $A should be installed. If not, then we can fail the test.

Wrap WebDriver script executions to check for errors upon termination of
the test script.

Look for the auraErrorMsg after loading the document as a potential
source of information for why the FW failed to load.

@rev goliver, esalman@
@bug W-1392193@