Implements (but does not enable) parallel test runs.
Introduce a @ThreadHostileTest annotation to mark some test suites as
unsuitable for parallel runs as a stop gap against hard to fix
dependencies on exclusive control over global state.

Introduce the notion of a 'descriptorName' which is simply the qualified
name without the pseudo protocol:// portion.

Rework usage of StringSourceLoader such that no test can assume a
static name in the string: namespace to avoid any collissions in the
StringSourceLoader.

The test executor threadpool is still hardcoded to 1 as we need to get a
better story around launching multiple local chromedrivers before making
parallel runs the default execution state.

@bug W-1319186@
@rev dpletter gtorres, rjayramappa@