Testfix for InputMultiSelect
Test was failing but after looking at the test, found out it doesn't check what the name/comments describe so removing it.  Looks like it was expecting an invalid value of an inputSelect to be added as a new option, but according to https://gus.salesforce.com/a07B0000000MmcbIAC it is not appropriate behavior for the base inputSelect.

@bug W-1455387@

@rev rjayaramappa@