AppCache test fix
The progres bar was being displayed and dismissed too fast for our tests to pick it up. Remove the UI test, and add a
js component test that checks that the correct AppCache events are recieved in the right order. Since
AppCacheProgressBarUITest.java tests that the right events display or dismiss the progress bar on the browser, and now
appCacheEventsTest.js tests we actually get these events, we should have confidence the progress bar is dispalyed to
the user.

@bug W-1570969@

@rev gtorres@