Remove (unused) service tests
The service tests have been disabled for well over a year, and seem to provide little benefit for the cost of re-enabling them. The tests take 6-7 minutes to run locally and only increase our line coverage by about 1.5%. More than half of this increase in code coverage is from the overloaded methods in LocalizationService (we have 7 formatDate methods for example), which can easily be tested in the Impl test class.

@bug W-1478190@

@rev rjayaramappa@