Updating to new Closure, which also uncovered error handling failures in GenerateJavascript and its use of DirectiveBasedJavascriptGroup: errors in the spawned subthreads were, in effect, ignored instead of stopping our build or other clients requesting generation.
The new closure-compiler, among other bugfixes, is nicer about not needing
finalizers, which may help our memory consumption woes.

The new CompositeRuntimeException is an analog of
java.concurrent.ExecutionException, except that it allows for multiple errors.
It will only be "caused by" the first error, but will carry the messages of all.

@rev goliver@
@bug W-1651249@