Safari testfix: disable websql test
Currently the websql database adapter hardcodes the amount of space it wants to use. This is greater than Safari's default 5mb limit so we get a pop up which we can't accept or override from the test. Disabling with attached bug until websql uses the user passed in config to calculate database size.

@bug W-1849610@

@rev rjayaramappa@