Preventing SimpleValue from nesting, with associated ValueProvider API change.
The nesting turned out to be positively used by LabelValueProvider, which would create
a placeholder SimpleValue, nest it inside a component attribute SV, and then update it
later when the "real" value had fetched.  So, instead, we rearrange Global... and
LabelValueProvider to use a setValue(expr, dest) API instead of returning a
placeholder.

As of this commit, AppCacheResourcesUITest has a few failures to untangle.

@rev bobyao@