Allows chaining off of a promise that is already resolved. Previously, a promise chained to an already executed promise would never be executed.
Allows promise chain insertion when a promise that is a member of a chain returns a new promise. Previously the returned promise would not be added to the chain and would not be executed.

@rev skrenzel@
@rev dmcmillen@
@bug W-2217782@

Code Review feedback 1

===

Allows chaining off of a promise that is already resolved. Previously, a promise chained to an already executed promise would never be executed.

Allows promise chain insertion when a promise that is a member of a chain returns a new promise. Previously the returned promise would not be added to the chain and would not be executed.

@rev skrenzel@
@rev dmcmillen@
@bug W-2217782@

Code Review feedback 1

===

Allows chaining off of a promise that is already resolved. Previously, a promise chained to an already executed promise would never be executed.

Allows promise chain insertion when a promise that is a member of a chain returns a new promise. Previously the returned promise would not be added to the chain and would not be executed.

@rev skrenzel@
@rev dmcmillen@
@bug W-2217782@