Exclude AuraId from coverage report
This generated file is not being picked up by Cobertura. Ignore the file
from coverage since it is just a giant enum for all our components and
attributes and is unrealistic to expect coverage for every enum entry.

AuraIdTest exists as a basic check and is currently running in
autobuilds.

@bug W-1756877@
@rev m.miller@