Add parameter to Test.clickOrTouch to bubble event
This changes default behavior to not bubble a touch event. Users can now
set a paremeter to allow bubbling to occur. Ideally, bubbling would be
enabled by default, but due to a bug in htmlHelper's FastClick and the
fact that we do not attach coordinates to touch events in tests cause
several test failures.

@bug W-2332239@
@rev mkohanfars@