Add separate wait for $A in WebDriver tests
Currently there isn't a way to tell if the a WebDriver test timed out
because $A was never loaded or if $A.finishedInit was never set to true
since they are both checked in the same wait. Separating out the two
checks will enable us to be sure where the test failed.

@rev mkohanfars@
@bug W-2361023@