Additional features for panels
Add an attribute to ui:panelType named “returnFocusElement” that allows
a consumer to specify the exact HTMLElement to return focus to when the
panel is closed. If the attribute is not specified, the default
behavior will be applied. The default is to use document.activeElement.

Change invocations to ui:panel controller functions to use the concrete
component so that components that extend ui:panel can have overridden
functions called.

Added an additional attribute flag to ui:panelType method “close” that
allows a consumer to specify whether the focus should return to a
specified element or not.

@rev stephen.woods@
@bug W-2981016@