ui-modal-panel-return-focus-fix
Fixes ui:modal so that it passes the shouldReturnFocus parameter
through to the ui:notify event.

Also fixes the panelManager2Helper.destroyPanel method so that
shouldReturnFocus defaults to true if it is not passed in. The
returnFocus element is now set AFTER setting the active panel because
setting the active panel would steal the focus.
@rev stephen.woods@