Unadaptable DirectiveBasedJavascriptGroupTest (#1171)
Depending on available classpath when ran, engine.js code will be included to gold file results. However, we purposely do not want to include engine.js in the gold file because it will change often. The intent is to test the formatting of the provided javascript in the tests. Not inclusion of engine.js

@bug W-3894613@