W-4424633 Fix tests to use a currency pattern for formatting currency. (#2133)
This fixes the test failures W-4424621, W-4424633, W-4555804.

After a back end change, now in Core, $A.get('$Locale.currencyFormat') returns "¤#,##0.00;(¤#,##0.00)". This causes the tests to fail that rely on the formatting to be "¤#,##0.00". Hence mock the format to be "¤#,##0.00" in tests so that they pass in Local as well as Core.