W-4730990 Modify reifyActions to return the newToken in case of invalid session error (#2506)
* Fix hydration test after namespace rename
"nativetest" namespace was renamed to "auranativetest" in fe61d26
* W-4667664 Modify reifyActions() for framework-level errors
AuraClientService.decode() handles a variety of framework-level errors that involve parsing of the Servlet's response payload. Native Priming requires this parsed result to be returned from the Framework.
Currently the HTTP status was hard-coded to 200. To allow us to properly invoke the decode() method for all scenarios, we must first supply the actual HTTP response code received by Native comms.

Thus, refactor reifyActions() to:
(1) allow the caller to specify the actual HTTP response code
(2) return the decode() method's result object when framework-level errors are encountered

* Fix hydration test after namespace rename
"nativetest" namespace was renamed to "auranativetest" in fe61d26
* W-4730990 Modify reifyActions to return error data with newToken in case of INVALID SESSION
* Adding hydration app test for error with no token
* Check for valid and new token in invalid session error
* Fix the !== of newToekn check