@W-5473871 The locale is properly read from the request prior to initializing the localization GVP (#3511)
Lazy load the GVPs in the `AuraContext` so they are initialized the first time the getter is called for them. This needs to be done after all of the other properties have been set via the context constructor. This is because the localization GVP internally reads the locale from the `AuraContext`. The locale is pulled form the request and we do not want to expose the request to wider scope than is absolutely necessary.
* Moved the calls for the GVPs to the `GlobalValueProviderFactory`
* Added additional Unit tests

@bug W-5473871@
@p4cl 16695075@
@rev linman@